+++ functions/Remove-PasswordStateResource.ps1
@@ -1,4 +1,4 @@
-<#
+﻿<#
 .SYNOPSIS
    A function to simplify the deletion of password state resources via the rest API
 .DESCRIPTION
@@ -30,7 +30,8 @@ function Remove-PasswordStateResource {
 
     begin {
         # Force TLS 1.2
-        [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
+        $SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol
+        [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.SecurityProtocolType]::Tls12
         # Import the environment
         $passwordstateenvironment = $(Get-PasswordStateEnvironment)
         # If the apikey is windowsauth then rebuild the uri string to match the windows auth apis, otherwise just build the api headers.
@@ -43,6 +44,7 @@ function Remove-PasswordStateResource {
             }
             APIKey {
                 $headers = @{"APIKey" = "$($passwordstateenvironment.Apikey)"}
+
             }
         }
     }
@@ -54,11 +56,23 @@ function Remove-PasswordStateResource {
             "Method"          = $method.ToUpper()
             "ContentType"     = $ContentType
         }
-        if ($extraparams) {
+        if (!$body){
+            $params.Remove("Body")
+        }
+        if ($headers -and $null -ne $extraparams.Headers) {
+            Write-Verbose "[$(Get-Date -format G)] Adding API Headers and extra param headers"
+            $headers += $extraparams.headers
+            $params += @{"headers" = $headers}
+            $skipheaders = $true
+        }
+        if ($extraparams -and $null -eq $extraparams.Headers){
+            Write-Verbose "[$(Get-Date -format G)] Adding extra parameter $($extraparams.keys) $($extraparams.values)"
             $params += $extraparams
         }
-        if ($headers) {
-            $params += $headers
+
+        if ($headers -and $skipheaders -ne $true) {
+            Write-Verbose "[$(Get-Date -format G)] Adding API Headers only"
+            $params += @{"headers" = $headers}
         }
         if ($PSCmdlet.ShouldProcess("[$($params.Method)] uri:$($params.uri) Headers:$($headers) Body:$($params.body)")) {
             Switch ($passwordstateenvironment.AuthType) {
@@ -81,6 +95,7 @@ function Remove-PasswordStateResource {
     }
 
     end {
+	    [System.Net.ServicePointManager]::SecurityProtocol = $SecurityProtocol
         return $result
     }
 }