+++ functions/Set-PasswordStateResource.ps1
@@ -1,4 +1,4 @@
-<#
+﻿<#
 .SYNOPSIS
    A function to simplify the modification/updates of password state resources via the rest API
 .DESCRIPTION
@@ -32,7 +32,9 @@ function Set-PasswordStateResource {
     )
 
     begin {
-        [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
+        # Force TLS 1.2
+        $SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol
+        [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.SecurityProtocolType]::Tls12
         $passwordstateenvironment = $(Get-PasswordStateEnvironment)
         # If the apikey is windowsauth then rebuild the uri string to match the windows auth apis, otherwise just build the api headers.
         Switch ($passwordstateenvironment.AuthType) {
@@ -56,11 +58,23 @@ function Set-PasswordStateResource {
             "ContentType"     = $ContentType
             "Body"            = $body
         }
-        if ($extraparams) {
+        if (!$body){
+            $params.Remove("Body")
+        }
+        if ($headers -and $null -ne $extraparams.Headers) {
+            Write-Verbose "[$(Get-Date -format G)] Adding API Headers and extra param headers"
+            $headers += $extraparams.headers
+            $params += @{"headers" = $headers}
+            $skipheaders = $true
+        }
+        if ($extraparams -and $null -eq $extraparams.Headers){
+            Write-Verbose "[$(Get-Date -format G)] Adding extra parameter $($extraparams.keys) $($extraparams.values)"
             $params += $extraparams
         }
-        if ($headers) {
-            $params += $headers
+
+        if ($headers -and $skipheaders -ne $true) {
+            Write-Verbose "[$(Get-Date -format G)] Adding API Headers only"
+            $params += @{"headers" = $headers}
         }
         if ($PSCmdlet.ShouldProcess("[$($params.Method)] uri:$($params.uri) Headers:$($headers) Body:$($params.body)")) {
             Switch ($passwordstateenvironment.AuthType) {
@@ -83,6 +97,7 @@ function Set-PasswordStateResource {
     }
 
     end {
+	    [System.Net.ServicePointManager]::SecurityProtocol = $SecurityProtocol
         return $result
     }
 }