+++ functions/Update-PasswordStatePassword.ps1
@@ -35,6 +35,7 @@
 function Update-PasswordStatePassword {
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingPlainTextForPassword', '', Justification = 'API requires password be passed as plain text')]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingUserNameAndPassWordParams', '', Justification = 'API requires password be passed as plain text')]
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidGlobalVars', '', Justification = 'Needed for backward compatability')]
     [CmdletBinding(SupportsShouldProcess = $true)]
     param (
         [parameter(Position = 0, ValueFromPipelineByPropertyName, Mandatory = $true)][int32]$passwordID,
@@ -45,20 +46,34 @@ function Update-PasswordStatePassword {
         [parameter(Position = 5, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$hostname,
         [parameter(Position = 6, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$notes,
         [parameter(Position = 7, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$url,
-        [parameter(Position = 8, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$reason
+        [parameter(Position = 8, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$reason,
+        [Parameter(Position = 9, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField1,
+        [Parameter(Position = 10, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField2,
+        [Parameter(Position = 11, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField3,
+        [Parameter(Position = 12, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField4,
+        [Parameter(Position = 13, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField5,
+        [Parameter(Position = 14, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField6,
+        [Parameter(Position = 15, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField7,
+        [Parameter(Position = 16, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField8,
+        [Parameter(Position = 17, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField9,
+        [Parameter(Position = 18, ValueFromPipelineByPropertyName, Mandatory = $false)][string]$GenericField10,
+        [parameter(Position = 19, ValueFromPipelineByPropertyName, Mandatory = $false)][switch]$PreventAuditing
+
     )
 
     begin {
-
+        . "$(Get-NativePath -PathAsStringArray "$PSScriptroot","PasswordStateClass.ps1")"
     }
 
     process {
-        if ($reason) {
+        If ($Reason) {
             $headerreason = @{"Reason" = "$reason"}
+            $parms = @{ExtraParams = @{"Headers" = $headerreason}}
         }
+        Else {$parms = @{}}
         if ($passwordID) {
             try {
-                $result = Find-PasswordStatePassword -PasswordID $passwordID -ErrorAction Stop
+                $result = Get-PasswordStatePassword -PasswordID $passwordID -ErrorAction Stop
                 Write-Verbose "[$(Get-Date -format G)] updating $($result.title)"
             }
             Catch {
@@ -74,7 +89,12 @@ function Update-PasswordStatePassword {
                 # Replace Result property with that of the bound parameter
                 $notprocess = "reason", "verbose", "erroraction", "debug", "whatif", "confirm"
                 if ($notprocess -notcontains $i) {
-                    $result.$($i) = $PSBoundParameters.$($i)
+                    if ($i -eq "Password" -and $PSBoundParameters.$($i).Gettype().Name -eq "EncryptedPassword"){
+                        $result.$($i) = $result.GetPassword()
+                    }
+                    Else {
+                        $result.$($i) = $PSBoundParameters.$($i)
+                    }
                 }
             }
             # Store in a new variable and remove all null values as password state doesn't like nulls.
@@ -92,12 +112,22 @@ function Update-PasswordStatePassword {
             # Update body variable to contain only the properties with data.
             $body = $body | Select-Object $selections
             # Write back to password state.
-            $output = Set-PasswordStateResource -uri "/api/passwords" -body "$($body|convertto-json)" -extraparams @{"Headers" = $headerreason}
+            $uri = "/api/passwords"
+            If ($PreventAuditing) {$uri += "PreventAuditing=$([System.Web.HttpUtility]::UrlEncode('True'))&"}
+            [PasswordResult]$output = Set-PasswordStateResource -uri $uri  -body "$($body|convertto-json)" @parms
+            foreach ($i in $output){
+                $i.Password = [EncryptedPassword]$i.Password
+            }
         }
 
     }
 
     end {
-        return $output
+        switch ($global:PasswordStateShowPasswordsPlainText) {
+            True {
+                $output.DecryptPassword()
+            }
+        }
+        Return $output
     }
 }